requirements.txt
# Web Framework & API
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20

# LangChain Ecosystem (v1.2+ Core)
langchain>=1.2.13
langchain-openai>=1.1.12
langchain-community>=0.4.1
langchain-text-splitters>=1.1.1
langchain-core>=1.2.22

# Vector Database & Processing
chromadb>=0.6.0
pypdf>=5.1.0

# Frontend & Environment
streamlit>=1.42.0
python-dotenv>=1.0.1